# Optimize InputSource handling with Cow and add benchmarks