--- a/benches/src/extract.rs
+++ b/benches/src/extract.rs
@@ -1,6 +1,6 @@
 use criterion::{Criterion, criterion_group, criterion_main};
-use lychee_lib::InputContent;
 use lychee_lib::extract::Extractor;
+use lychee_lib::{FileType, InputContent};
 use std::hint::black_box;
 use std::path::PathBuf;
 
@@ -11,6 +11,29 @@ fn extract(paths: &[PathBuf]) {
         let extracted = extractor.extract(&content);
         println!("{}", extracted.len());
     }
+}
+
+fn benchmark_input_content_creation(c: &mut Criterion) {
+    let test_data = "https://example.com/link1 https://example.com/link2 https://example.com/link3"
+        .repeat(1000);
+    let owned_string = test_data.clone();
+
+    c.bench_function("InputContent::from_string", |b| {
+        b.iter(|| InputContent::from_string(black_box(&test_data), FileType::Markdown))
+    });
+
+    c.bench_function("InputContent::from_owned_string (optimized)", |b| {
+        b.iter(|| {
+            InputContent::from_owned_string(black_box(owned_string.clone()), FileType::Markdown)
+        })
+    });
+
+    c.bench_function("InputContent::from_string with owned (baseline)", |b| {
+        b.iter(|| {
+            let owned_string = test_data.clone();
+            InputContent::from_string(black_box(&owned_string), FileType::Markdown)
+        })
+    });
 }
 
 fn benchmark(c: &mut Criterion) {
@@ -28,6 +51,6 @@ fn benchmark(c: &mut Criterion) {
 criterion_group!(
     name = benches;
     config = Criterion::default().sample_size(10);
-    targets = benchmark
+    targets = benchmark, benchmark_input_content_creation
 );
 criterion_main!(benches);