--- a/lychee-lib/src/types/input/resolver.rs
+++ b/lychee-lib/src/types/input/resolver.rs
@@ -152,7 +152,7 @@ impl InputResolver {
                     yield ResolvedInputSource::Stdin;
                 },
                 InputSource::String(s) => {
-                    yield ResolvedInputSource::String(s.clone());
+                    yield ResolvedInputSource::String(s.to_string());
                 }
             }
         }