Out of curiosity, I finally tackled this TODO from the code:

```rust
// TODO: consider using Cow (to avoid one .clone() for String types)
```

I took a look and it turns out it doesn't touch as many places as I thought it would, so I went with it.

The result is lesser allocations for places where we handle raw input strings.

I did add a micro-benchmark for input content creation and it turns out to be decently fast. 

On my machine:

  - Extract from large docs: 28% improvement (21.4ms vs 29.3ms baseline)
  - InputContent::from_string: 28% improvement (5.3µs vs 7.4µs baseline)
  - InputContent::from_owned_string: 9% improvement (14.2µs vs 15.7µs baseline)
  - from_string with owned baseline: 14% improvement (8.1µs vs 9.4µs baseline)

Not earth-shaking, but we're already operating at a decent level of performance by now, so I take it.

The question is: what do y'all think? Does it make the code too complex or is the complexity manageable? Feedback very welcome!